move truncate to attrnorm package
